Add `cucumber:mongo` script
Following @appinteractive's suggestions

`|| sleep infinity` here will make sure the command for starting mongo
won't terminate the parallel execution of `concurrently`. That would be
the case if your mongo is already running, e.g. because of a system-wide
mongodb daemon. But we don't want to terminate the commmand because that
would in turn terminate `concurrently`. Instead we want the last command
`cucumber-js` to exit first.