Simplify package.json, remove dev:local script
The script was assuming a POSIX shell and a command `ifconfig` available
on the system. I don't have either. I guess we can just remove it.

You can access the server from the local network if you override `host`
in your `config/local.json`. This is much better in my opinion.