Fix usersettings hook if data === undefined
When I run the seed data, I get several error messages when I want to
delete my account. The error that is fixed with this commit is
```
Cannot read property 'blacklist' of undefined
```

Another error which pops up is
```
Cannot read property 'Forbidden' of undefined
```
This error is **NOT** resolved with this commit. Fortunately, some
resources get deleted when the user pushes the button, so all you have
to do is to click the "delete account" button multiple times.